move account parsing to its own module
